+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWYNGDGJK7FJXGQMYM2Z10ZT"
+        "01KX05Z72MZVCGW5DBGRHF2MAQ"
     )
     r"""The unique identifier of the data source"""