+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-07T16:09:54.663Z"
+    created: Optional[str] = "2026-07-08T06:16:48.202Z"
 
-    updated: Optional[str] = "2026-07-07T16:09:54.663Z"
+    updated: Optional[str] = "2026-07-08T06:16:48.202Z"
 
     updated_by_id: OptionalNullable[str] = UNSET