+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-07T16:09:54.663Z"
+    created: Optional[str] = "2026-07-08T06:16:48.202Z"
 
-    updated: Optional[str] = "2026-07-07T16:09:54.663Z"
+    updated: Optional[str] = "2026-07-08T06:16:48.202Z"
 
     updated_by_id: OptionalNullable[str] = UNSET