+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsepython.py
@@ -40,9 +40,9 @@ class EvaluatorResponsePython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-07T16:09:54.663Z"
+    created: Optional[str] = "2026-07-08T06:16:48.202Z"
 
-    updated: Optional[str] = "2026-07-07T16:09:54.663Z"
+    updated: Optional[str] = "2026-07-08T06:16:48.202Z"
 
     updated_by_id: OptionalNullable[str] = UNSET