+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWYNGDGN2SBDJ2YWJ261J4F6"
+        "01KX05Z72Q93PVYE3GXNSEVN5G"
     )
     r"""The unique identifier of the data source"""