+++ packages/orq-rc/RELEASES.md
@@ -9329,3 +9329,13 @@ Based on:
 - [python v4.12.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.5 - packages/orq-rc
+
+## 2026-07-08 06:19:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.0 (2.916.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.6 - packages/orq-rc